🐛 Fix single pauli measurement issue (#17)
* :bug: Fix measurement handling for single pauli instructions

Previously, the `calculate_measurement_type` method required the measured observables to be a list, which is the case whenever there are multiple Pauli operators as an observable, i.e., `return qml.expval(qml.PauliX(0) @ qml.PauliZ(1))`. However, issues arise when there is just a single operator, which hasn't been converted into a list. 
This patch handles single observables as lists with a single element, which fixes the issue.